Unification, now search slot is not virtual, but Dictionary::quickSearch and Dictionary::subsequentSearch are pure virtual methods now.
Because search algorithms are working with assumed ASCII strings it's not effective to read files with Qt methods and it is not effective to store ASCII strings in QString objects.

With that said, I switched from QFile and QTextStream to std::filesystem and std::ifstream.
But it requires additional testing, as it is not very stable if works with non-ASCII characters.